pkg/helm/client: refactoring from pkg/helm
