+++ helm-app-operator/cmd/manager/main.go
@@ -29,6 +29,7 @@ import (
 	"sigs.k8s.io/controller-runtime/pkg/runtime/signals"
 
 	"github.com/operator-framework/helm-app-operator-kit/helm-app-operator/pkg/helm"
+	"github.com/operator-framework/helm-app-operator-kit/helm-app-operator/pkg/helm/client"
 	"github.com/operator-framework/helm-app-operator-kit/helm-app-operator/pkg/helm/controller"
 )
 
@@ -64,7 +65,7 @@ func main() {
 
 	// Create Tiller's storage backend and kubernetes client
 	storageBackend := storage.Init(driver.NewMemory())
-	tillerKubeClient, err := helm.NewTillerClientFromManager(mgr)
+	tillerKubeClient, err := client.NewFromManager(mgr)
 	if err != nil {
 		logrus.Fatal(err)
 	}
